requirements.txt
pymysql==1.1.2
cryptography==46.0.2
python-dotenv==1.0.0
